+++ addon/components/file-browser/component.js
@@ -2,7 +2,11 @@ import Ember from 'ember';
 import layout from './template';
 
 import loadAll from 'ember-osf/utils/load-relationship';
+<<<<<<< HEAD
 import outsideClick from 'ember-osf/utils/outside-click';
+=======
+import Analytics from '../../mixins/analytics';
+>>>>>>> Add analytics to file items for quickfiles
 
 /**
  * File browser widget
@@ -13,7 +17,7 @@ import outsideClick from 'ember-osf/utils/outside-click';
  * ```
  * @class file-browser
  */
-export default Ember.Component.extend({
+export default Ember.Component.extend(Analytics, {
     // TODO: Improve documentation in the future
     layout,
     //Can be overwritten to have a trimmed down display, these are all the options available to be displayed
@@ -148,6 +152,7 @@ export default Ember.Component.extend({
             this.get('toast').error(response);
         },
         success(_, __, file, response) {
+            this.send('track', 'upload', 'track', 'Quick Files - Upload');
             this.get('uploading').removeObject(file);
             let data = response.data.attributes;
             //OPTIONS (some not researched)