+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - enable open on select
 - ability to specify a pre-selected file in file-browser
 - Stylesheet for the footer to match OSF styles
+- Analytics to `file-browser` and `file-browser-item` for Quick Files:
+  - Share
+  - Download
+  - Download as zip
+  - Upload
 
 ### Changed
 - getContents() function for files to use `redirect = true` and `mode = 'render'`